replace dotenv with vite import.meta.env
