+++ .env
@@ -1,3 +1,3 @@
-ALGOLIA_APP_ID=0OJ5UL9OJX
-ALGOLIA_SEARCH_KEY=63f563566cdd6de606e2bb0fdc291994
+VITE_ALGOLIA_APP_ID=0OJ5UL9OJX
+VITE_ALGOLIA_SEARCH_KEY=63f563566cdd6de606e2bb0fdc291994
 ALGOLIA_ADMIN_KEY=016cad360bc0f3dd0816ea6792644424
